requirements.txt
beautifulsoup4==4.11.1
httpx==0.23.3
numpy==1.24.0
opencv_python==4.7.0.68
pycryptodome==3.16.0
